+++ packages/components/CHANGELOG.md
@@ -18,7 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - React: `'use client';` directive is applied on all components for main and `ssr` sub-package
   ([#2654](https://github.com/porsche-design-system/porsche-design-system/pull/2654))
-- Breakpoint customizable property `column` to `Segmented Control` to control the width of items
+- Breakpoint customizable property `columns` to `Segmented Control` to control the width of items
   ([#2652](https://github.com/porsche-design-system/porsche-design-system/pull/2652))
 
 #### Fixed